[add] 판매상품별 기능 추가 했지만, 값 몇개 못불러옴

+++ BPS/src/main/java/com/sales/BPS/mproduct/service/VoucherService.java
@@ -76,8 +76,8 @@ public class VoucherService {
     // 매출액 계산 로직
     private void calculateVoucSales(Voucher voucher) {
         if (voucher.getVoucSale() != null && voucher.getVoucAmount() != null) {
-            Long sales = voucher.getVoucSale() * voucher.getVoucAmount();
-            voucher.setVoucSales(Math.toIntExact(sales));
+            Long sales = (long) (voucher.getVoucSale() * voucher.getVoucAmount());
+            voucher.setVoucSales(sales);
         }
     }
 }
